Moved model classes closely to their usages
